Moved component to separate file.
